+++ apps/frontend/src/modules/routes/libs/types/types.ts
@@ -1,4 +1,5 @@
 export { type PatchActionPayload } from "./patch-action-payload.type.js";
+export { type UploadImageActionPayload } from "./upload-image-action-payload.type.js";
 export {
 	type RouteCreateRequestDto,
 	type RouteFindAllOptions,