+++ apps/frontend/src/modules/routes/routes-api.ts
@@ -58,7 +58,7 @@ class RoutesApi extends BaseHTTPApi {
 		return await response.json();
 	}
 
-	public async getRouteById(
+	public async getById(
 		id: number,
 	): Promise<APIResponse<RouteGetByIdResponseDto>> {
 		const response = await this.load<APIResponse<RouteGetByIdResponseDto>>(
@@ -72,7 +72,7 @@ class RoutesApi extends BaseHTTPApi {
 		return await response.json();
 	}
 
-	public async patchRoute({
+	public async patch({
 		id,
 		payload,
 	}: PatchActionPayload): Promise<APIResponse<RouteGetByIdResponseDto>> {