+++ packages/shared/src/modules/points-of-interest/libs/types/points-of-interest-get-all-query.type.ts
@@ -1,11 +1,10 @@
 import { type PaginationQuery } from "../../../../libs/types/types.js";
 
-type PointsOfInterestQueryRequest = PaginationQuery & {
+type PointsOfInterestGetAllQuery = PaginationQuery & {
 	latitude?: number;
 	longitude?: number;
-	name?: string;
 	radius?: number;
 	search?: string;
 };
 
-export { type PointsOfInterestQueryRequest };
+export { type PointsOfInterestGetAllQuery };